# backend/src/swagger.yaml
openapi: 3.0.3
info:
  title: Antioquia Natural API
  description: >
    API REST para la aplicación Antioquia Natural — Gobernación de Antioquia.
    Expone información sobre especies, familias, grupos taxonómicos, subregiones
    y recursos hídricos del departamento. También incluye los endpoints del panel
    de administración de curadores (JPL y Guarda Cuencas).
  version: 1.0.0
  contact:
    name: Sebastián Guzmán Díaz
    email: sguzmand@gmail.com
servers:
  - url: /api
    description: Servidor local / producción (relativo)
...
paths:
...
  /admin/login:
    post:
      tags: [Admin — Autenticación]
      summary: Iniciar sesión como curador
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          description: Contraseña incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/Error500'
...
  /admin/me:
    get:
      tags: [Admin — Autenticación]
      summary: Verificar sesión activa
      operationId: adminMe
      responses:
        '200':
          description: Sesión válida
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                    example: true
        '401':
          description: Sin sesión activa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Admin — JPL ───────────────────────────────────────────────────────────
  /admin/jpl/meses:
    get:
      tags: [Admin — JPL]
      summary: Listar meses disponibles en JPL
      operationId: jplListMeses
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Lista de meses con conteo de fotos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesResumen'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/jpl/fotos/{mes}:
    get:
      tags: [Admin — JPL]
      summary: Listar fotos JPL de un mes
      operationId: jplGetFotos
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
      responses:
        '200':
          description: Fotos del mes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JplFoto'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'

    post:
      tags: [Admin — JPL]
      summary: Subir nueva foto JPL
      operationId: jplUploadFoto
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [foto]
              properties:
                foto:
                  type: string
                  format: binary
                  description: Archivo de imagen (JPG/PNG/WebP, máx 10 MB)
                usuario:
                  type: string
                municipio:
                  type: string
                grupo:
                  type: string
                  enum: [aves, anfibios_reptiles, mariposas, polillas, mamiferos, animales_domesticos, peces, orquideas, arboles_nativos]
                iucn:
                  type: string
                  enum: [LC, NT, VU, EN, CR, DD]
                endemica:
                  type: boolean
                comentario:
                  type: string
      responses:
        '201':
          description: Foto subida y convertida a WebP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JplFoto'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/jpl/fotos/{mes}/{id}:
    put:
      tags: [Admin — JPL]
      summary: Actualizar metadatos de una foto JPL
      operationId: jplUpdateFoto
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
        - $ref: '#/components/parameters/FotoIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JplFotoUpdate'
      responses:
        '200':
          description: Foto actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JplFoto'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/jpl/fotos/{id}:
    delete:
      tags: [Admin — JPL]
      summary: Eliminar una foto JPL
      operationId: jplDeleteFoto
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/FotoIdParam'
      responses:
        '200':
          description: Foto eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/jpl/publicar/{mes}:
    post:
      tags: [Admin — JPL]
      summary: Publicar JSON mensual de JPL
      operationId: jplPublicar
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
      responses:
        '200':
          description: JSON publicado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  archivo:
                    type: string
                    example: fotos_2026_06.json
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'

  # ─── Admin — Guarda Cuencas ────────────────────────────────────────────────
  /admin/gc/meses:
    get:
      tags: [Admin — Guarda Cuencas]
      summary: Listar meses disponibles en Guarda Cuencas
      operationId: gcListMeses
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Lista de meses con conteo de fotos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesResumen'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/gc/fotos/{mes}:
    get:
      tags: [Admin — Guarda Cuencas]
      summary: Listar fotos GC de un mes
      operationId: gcGetFotos
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
      responses:
        '200':
          description: Fotos del mes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GcFoto'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'

    post:
      tags: [Admin — Guarda Cuencas]
      summary: Subir nueva foto GC
      operationId: gcUploadFoto
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [foto]
              properties:
                foto:
                  type: string
                  format: binary
                  description: Archivo de imagen (JPG/PNG/WebP, máx 10 MB)
                cuenca:
                  type: string
                subregion:
                  type: string
                municipio:
                  type: string
                comentario:
                  type: string
      responses:
        '201':
          description: Foto subida y convertida a WebP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GcFoto'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/gc/fotos/{mes}/{id}:
    put:
      tags: [Admin — Guarda Cuencas]
      summary: Actualizar metadatos de una foto GC
      operationId: gcUpdateFoto
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MesParam'
        - $ref: '#/components/parameters/FotoIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GcFotoUpdate'
      responses:
        '200':
          description: Foto actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GcFoto'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/gc/fotos/{id}:
    delete:
      tags: [Admin — Guarda Cuencas]
      summary: Eliminar una foto GC
      operationId: gcDeleteFoto
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/FotoIdParam'
      responses:
        '200':
          description: Foto eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Cookie de sesión obtenida al hacer POST /api/admin/login
  parameters:
    MesParam:
      name: mes
      in: path
      required: true
      description: Mes en formato YYYY_MM
      schema:
        type: string
        pattern: '^\d{4}_\d{2}$'
        example: 2026_06
    FotoIdParam:
      name: id
      in: path
      required: true
      description: ID MongoDB de la foto
      schema:
        type: string
        example: 6650f1a2b3c4d5e6f7a8b9c0
  responses:
    Error401:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error500:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Recurso no encontrado
...
    MesResumen:
      type: object
      properties:
        mes:
          type: string
          example: 2026_06
        totalFotos:
          type: integer
          example: 42
...
    JplFotoUpdate:
      type: object
      properties:
        usuario:
          type: string
        municipio:
          type: string
        grupo:
          type: string
        iucn:
          type: string
        endemica:
          type: boolean
        comentario:
          type: string
...